authorization implemented

+++ src/api/notes/handler.js
@@ -13,7 +13,15 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
-    const noteId = await this._service.addNote({ title, body, tags });
+    const { id: credentialId } = request.auth.credentials;
+
+    const noteId = await this._service.addNote({
+      title,
+      body,
+      tags,
+      owner: credentialId,
+    });
+
     const response = h.response({
       status: 'success',
       message: 'Catatan berhasil ditambahkan',
@@ -25,8 +33,9 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -35,8 +44,11 @@ class NotesHandler {
     };
   }
 
-  async getNoteByIdHandler(request) {
+  async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
     return {
       status: 'success',
@@ -46,10 +58,12 @@ class NotesHandler {
     };
   }
 
-  putNoteByIdHandler(request) {
+  async putNoteByIdHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
+    await this._service.verifyNoteOwner(id, credentialId);
     this._service.editNoteById(id, request.payload);
 
     return {
@@ -58,9 +72,11 @@ class NotesHandler {
     };
   }
 
-  deleteNoteByIdHandler(request) {
+  async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
+    await this._service.verifyNoteOwner(id, credentialId);
     this._service.deleteNoteById(id);
 
     return {
